+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Next Release
+
+- Improve performance of fetching public keys by storing them in an ETS table
+
 ## 0.1.0 Initial Release
 
 - 🔥